redesigned Profile page
