--- src/components/ui/Loading/Loading.module.scss
@@ -1,18 +1,0 @@
-.loading {    
-    border: 8px solid $text-secondary;
-    border-top: 8px solid $color-primary;
-    border-radius: 50%;
-    width: 80px;
-    height: 80px;
-    animation: spin 2s linear infinite;
-  
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-  
-  @keyframes spin {
-    0% { transform: translate(-50%, -50%) rotate(0deg); }
-    100% { transform: translate(-50%, -50%) rotate(360deg); }
-  }
