--- src/components/ui/Loading/index.tsx
@@ -1,11 +1,0 @@
-import styles from './Loading.module.scss'
-
-const Loading = () => {
-    return (
-        <div className={styles.loading__container}>
-            <div className={styles.loading}></div>
-        </div>
-    )
-}
-
-export default Loading;
