--- src/routes/Profile/Profile.module.scss
@@ -1,40 +1,0 @@
-.profile {
-    &__title {
-        margin-top: 50px;
-    }
-
-    &__info {
-        margin-top: 50px;
-        display: flex;
-        align-items: center;
-
-    }
-    &__watched {
-        display: flex;
-        flex-direction: column;
-        gap: 20px;
-        margin-left: 20px;
-
-        &_item {
-            display: flex;
-            align-items: center;
-            gap: 10px;
-        }
-    }
-
-    &__button {
-        font-size: $fz-small;
-        color: $text-primary;
-        padding: 10px;
-        background-color: $color-primary;
-        border-radius: 10px;
-    }
-
-    &__count {
-        font-size: $fz-small;
-        color: $text-primary;
-        padding: 10px;
-        background-color: $text-tertiary;
-        border-radius: 10px;
-    }
-}
